src/api/services/calendar_service.py
import google.auth
from googleapiclient.discovery import build


def fetch_events():
    # Logic to fetch upcoming events from Google Calendar
    # You'll need to use the Google Calendar API here with proper authentication

    # Example code (you'll need to configure authentication)
    credentials, _ = google.auth.default()
    service = build('calendar', 'v3', credentials=credentials)
    calendar_id = 'primary'
    events_result = service.events().list(calendarId=calendar_id).execute()
    events = events_result.get('items', [])

    return events
